register user making

+++ src/utils/Cloudinary.js
@@ -2,7 +2,7 @@ import { v2 as cloudinary } from 'cloudinary';
 import fs from"fs";
 
 cloudinary.config({ 
-        cloud_name:process.env.CLOUDIANRY_NAME, 
+        cloud_name:process.env.CLOUDINARY_NAME, 
         api_key: process.env.CLOUDINARY_API_KEY, 
         api_secret: process.env.CLOUDINARY_API_SECRET
 });
@@ -13,13 +13,14 @@ const uploadOnCloudinary= async (localFile)=>{
         const response=await cloudinary.uploader.upload(localFile,{
             resource_type:"auto"
         });
-        console.log("file is upload in cloudinary",response.url);
+        fs.unlinkSync(localFile)
         return response;
     }
     catch(error){
-        fs.unlinkSync(localFile);
+        if (fs.existsSync(localFile)) {
+            fs.unlinkSync(localFile);
+        }
         return null;
-
     }
 }
